+++ src/transport/Bus.java
@@ -1,0 +1,14 @@
+package transport;
+
+public class Bus extends Transport {
+
+    public Bus(String brand,
+               String model,
+               int year,
+               String country,
+               String color,
+               int maxSpeed) {
+        super(brand,model,year,country,color,maxSpeed);
+    }
+
+}
